api-customer/pyproject.toml
[tool.poetry]
name = "api-auth"
version = "0.1.0"
description = ""
authors = ["Davi Lucciola <davilucciola14@gmail.com>"]
readme = "../README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.10.6"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
alembic = "^1.11.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
ormar = {extras = ["sqlite"], version = "^0.12.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
